requirements.txt
Flask==3.1.3
Flask_Cors==4.0.0
python-dotenv==1.2.2
resend==2.27.0
stripe==15.0.1
gunicorn==25.3.0
google-auth==2.32.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.194.0